requirements.txt
Flask==2.3.2
flask-cors==3.0.10
flask-sqlalchemy==3.0.5
flask-bcrypt==1.0.1
flask-jwt-extended==4.4.4
